new success logic added

+++ Jenkinsfile
@@ -1,6 +1,7 @@
 node {
 
     def success = false
+    def deployed = false
     def app
     def image = 'registry.hub.docker.com/jnnngs/nginx-helloworld'
     def buildtag = "nginx-hello-world-${env.BUILD_NUMBER}"
@@ -10,9 +11,8 @@ node {
 
     notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>Building</b>: ${env.JOB_NAME}, <b>Build</b> #${env.BUILD_NUMBER}, <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN
     
-    stage "Checkout" {
+    stage "Checkout" 
         checkout scm
-    }
 
     stage " Creating docker image"
         sh "echo '[i] building docker image'"
@@ -29,38 +29,52 @@ node {
             sh "docker ps | grep 'nginx-hw-example-${env.BUILD_NUMBER}'"
             echo "Docker Availability GOOD"
             notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>Testing</b>: <a target='_blank' href='${env.JOB_NAME}'>${env.BUILD_TAG}</a>, <b>Test</b> #${env.BUILD_NUMBER}, <b>Container</b> EXISTS, <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN
+            deployed = true
             success = true
         } catch (Exception e) {
             echo "Docker Availability BAD"
             notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>SUCCESS set to FALSE</b>: Docker Availability BAD <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN 
             success = false
         }
-	if ( success ) { 
-        echo "Yes"
-    } else {
-        echo "No"
-    }         
-    stage "Push image to DockerHub"
-	try { 
-    	    docker.withRegistry('https://registry.hub.docker.com', 'docker-hub-credentials')
-		{
-                app.push("latest")  
-	        }
-	    echo "Push Success: TAG latest"
-	    notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>Dockerhub push</b>: <a target='_blank' href='${env.JOB_NAME}'>${env.BUILD_TAG}</a>, <b>DockerHub</b> #${env.BUILD_NUMBER}, <b>PUSH SUCCESS</b>, <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN 
-	} catch (Exception e) {
-            echo "Push failed"
-            notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>Dockerhub push</b>: <a target='_blank' href='${env.JOB_NAME}'>${env.BUILD_TAG}</a>, <b>DockerHub</b> #${env.BUILD_NUMBER}, <b>PUSH FAILED</b>, <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN 
-            sh "echo '[i] cleaning up all resources'"
-            sh "docker rm -f nginx-hw-example-${env.BUILD_NUMBER}"
-            sh "docker rmi ${buildtag}"
-	    return false
-        }		
-    
-    stage "finish build & clean-up"
+
+    if ( success ) {        
+        stage "Push image to DockerHub"
+        try { 
+                docker.withRegistry('https://registry.hub.docker.com', 'docker-hub-credentials')
+            {
+                    app.push("latest")  
+                }
+            echo "Push Success: TAG latest"
+            notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>Dockerhub push</b>: <a target='_blank' href='${env.JOB_NAME}'>${env.BUILD_TAG}</a>, <b>DockerHub</b> #${env.BUILD_NUMBER}, <b>PUSH SUCCESS</b>, <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN 
+            success = true
+        } catch (Exception e) {
+                echo "Push failed"
+                notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>Dockerhub push</b>: <a target='_blank' href='${env.JOB_NAME}'>${env.BUILD_TAG}</a>, <b>DockerHub</b> #${env.BUILD_NUMBER}, <b>PUSH FAILED</b>, <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN 
+                sh "echo '[i] cleaning up all resources'"
+                sh "docker rm -f nginx-hw-example-${env.BUILD_NUMBER}"
+                sh "docker rmi ${buildtag}"
+                notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>SUCCESS set to FALSE</b>: Dockerhub PUSH BAD <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN 
+                success = false
+            }	
+    }	
+     if ( delpoyed ) {     
+        stage "finish build & clean-up"
         sh "echo '[i] cleaning up all resources'"
         sh "docker rm -f nginx-hw-example-${env.BUILD_NUMBER}"
         sh "docker rmi ${buildtag}"
-        notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>Finished</b>: <a target='_blank' href='${env.JOB_NAME}'>${env.BUILD_TAG}</a>, <b>Build</b> #${env.BUILD_NUMBER}, <b>Status</b> ${currentBuild.currentResult}, <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN
+        deployed = false
+     }
+     stage "Check if finished successfully"
+     if ( success ) { 
+         echo "Finshed Successfully"
+         notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>Finished</b>: <a target='_blank' href='${env.JOB_NAME}'>${env.BUILD_TAG}</a>, <b>Build</b> #${env.BUILD_NUMBER}, <b>Status</b> ${currentBuild.currentResult}, <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN  
+         notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <a target='_blank' href='${env.BUILD_LOG}'>Build log</a>", token: env.SLACK_TOKEN
+         return true
+     } else {
+        echo "Finshed UnSuccessfully"
+        notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <b>Finished</b>: <a target='_blank' href='${env.JOB_NAME}'>${env.BUILD_TAG}</a>, <b>Build</b> #${env.BUILD_NUMBER}, <b>Status</b> ${currentBuild.currentResult}, <b>Duration</b> ${currentBuild.durationString.minus(' and counting')}", token: env.SLACK_TOKEN  
         notifyEvents message: "${new Date().format('dd MMM yyyy HH:mm:ss')} - <a target='_blank' href='${env.BUILD_LOG}'>Build log</a>", token: env.SLACK_TOKEN
+        return false
+     }
+     
 }
